[SDK]ExorAIO: Resharped.

+++ ExorAIO/AIO/Champions/Kalista/Properties/Utilities/Healthbars.cs
@@ -2,9 +2,9 @@ using System.Drawing;
 using System.Linq;
 using ExorAIO.Utilities;
 using LeagueSharp;
+using LeagueSharp.Data.Enumerations;
 using LeagueSharp.SDK;
 using LeagueSharp.SDK.UI;
-using LeagueSharp.Data.Enumerations;
 
 namespace ExorAIO.Champions.Kalista
 {
@@ -32,56 +32,71 @@ namespace ExorAIO.Champions.Kalista
                             h.IsValidTarget() &&
                             Bools.IsPerfectRendTarget(h) &&
                             (h is Obj_AI_Hero ||
-                            Vars.JungleList.Contains(h.CharData.BaseSkinName))).ToList().ForEach(unit =>
-                        {
-                            /// <summary>
-                            ///     Defines what HPBar Offsets it should display.
-                            /// </summary>
-                            var mobOffset = Vars.JungleHpBarOffsetList.FirstOrDefault(x => x.BaseSkinName.Equals(unit.CharData.BaseSkinName));
+                             Vars.JungleList.Contains(h.CharData.BaseSkinName))).ToList().ForEach(unit =>
+                             {
+                                 /// <summary>
+                                 ///     Defines what HPBar Offsets it should display.
+                                 /// </summary>
+                                 var mobOffset =
+                                     Vars.JungleHpBarOffsetList.FirstOrDefault(
+                                         x => x.BaseSkinName.Equals(unit.CharData.BaseSkinName));
 
-                            var width = (int)(Vars.JungleList.Contains(unit.CharData.BaseSkinName) ? mobOffset.Width : Vars.Width);
-                            var height = (int)(Vars.JungleList.Contains(unit.CharData.BaseSkinName) ? mobOffset.Height : Vars.Height);
-                            var xOffset = (int)(Vars.JungleList.Contains(unit.CharData.BaseSkinName) ? mobOffset.XOffset : Vars.XOffset);
-                            var yOffset = (int)(Vars.JungleList.Contains(unit.CharData.BaseSkinName) ? mobOffset.YOffset : Vars.YOffset);
+                                 var width = Vars.JungleList.Contains(unit.CharData.BaseSkinName)
+                                     ? mobOffset.Width
+                                     : Vars.Width;
+                                 var height = Vars.JungleList.Contains(unit.CharData.BaseSkinName)
+                                     ? mobOffset.Height
+                                     : Vars.Height;
+                                 var xOffset = Vars.JungleList.Contains(unit.CharData.BaseSkinName)
+                                     ? mobOffset.XOffset
+                                     : Vars.XOffset;
+                                 var yOffset = Vars.JungleList.Contains(unit.CharData.BaseSkinName)
+                                     ? mobOffset.YOffset
+                                     : Vars.YOffset;
 
-                            var barPos = unit.HPBarPosition;
-                            {
-                                barPos.X += xOffset;
-                                barPos.Y += yOffset;
-                            }
+                                 var barPos = unit.HPBarPosition;
+                                 {
+                                     barPos.X += xOffset;
+                                     barPos.Y += yOffset;
+                                 }
 
-                            var drawEndXPos = barPos.X + width * (unit.HealthPercent / 100);
-                            var drawStartXPos = barPos.X + (Vars.GetRealHealth(unit) >
-                                (float)GameObjects.Player.GetSpellDamage(unit, SpellSlot.E) +
-                                (float)GameObjects.Player.GetSpellDamage(unit, SpellSlot.E, DamageStage.Buff)
-                                    ? width * (((Vars.GetRealHealth(unit) -
-                                        ((float)GameObjects.Player.GetSpellDamage(unit, SpellSlot.E) +
-                                         (float)GameObjects.Player.GetSpellDamage(unit, SpellSlot.E, DamageStage.Buff))) / unit.MaxHealth * 100) / 100)
-                                    : 0);
+                                 var drawEndXPos = barPos.X + width*(unit.HealthPercent/100);
+                                 var drawStartXPos = barPos.X + (Vars.GetRealHealth(unit) >
+                                                                 (float)
+                                                                     GameObjects.Player.GetSpellDamage(unit, SpellSlot.E) +
+                                                                 (float)
+                                                                     GameObjects.Player.GetSpellDamage(unit, SpellSlot.E,
+                                                                         DamageStage.Buff)
+                                     ? width*((Vars.GetRealHealth(unit) -
+                                               ((float) GameObjects.Player.GetSpellDamage(unit, SpellSlot.E) +
+                                                (float)
+                                                    GameObjects.Player.GetSpellDamage(unit, SpellSlot.E,
+                                                        DamageStage.Buff)))/unit.MaxHealth*100/100)
+                                     : 0);
 
-                            Drawing.DrawLine(
-                                drawStartXPos,
-                                barPos.Y,
-                                drawEndXPos,
-                                barPos.Y,
-                                height,
-                                Vars.GetRealHealth(unit) <
-                                    (float)GameObjects.Player.GetSpellDamage(unit, SpellSlot.E) +
-                                    (float)GameObjects.Player.GetSpellDamage(unit, SpellSlot.E, DamageStage.Buff)
-                                        ? Color.Blue 
-                                        : Color.Orange
-                            );
+                                 Drawing.DrawLine(
+                                     drawStartXPos,
+                                     barPos.Y,
+                                     drawEndXPos,
+                                     barPos.Y,
+                                     height,
+                                     Vars.GetRealHealth(unit) <
+                                     (float) GameObjects.Player.GetSpellDamage(unit, SpellSlot.E) +
+                                     (float) GameObjects.Player.GetSpellDamage(unit, SpellSlot.E, DamageStage.Buff)
+                                         ? Color.Blue
+                                         : Color.Orange
+                                     );
 
-                            Drawing.DrawLine(
-                                drawStartXPos,
-                                barPos.Y,
-                                drawStartXPos,
-                                barPos.Y + height + 1,
-                                1,
-                                Color.Lime
-                            );
-                        }
-                    );
+                                 Drawing.DrawLine(
+                                     drawStartXPos,
+                                     barPos.Y,
+                                     drawStartXPos,
+                                     barPos.Y + height + 1,
+                                     1,
+                                     Color.Lime
+                                     );
+                             }
+                        );
                 }
             };
         }
